requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
boto3==1.35.0
botocore==1.35.99
certifi==2025.10.5
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
chardet==5.2.0
charset-normalizer==3.4.4
constantly==23.10.4
cryptography==45.0.7
daphne==4.2.1
deprecation==2.1.0
dj-database-url==3.0.1
Django==5.1.4
django-cors-headers==4.8.0
django-storages==1.14.4
django-tenants==3.7.0
djangorestframework==3.16.1
drf-nested-routers==0.95.0
exceptiongroup==1.3.0
Faker==37.6.0
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jmespath==1.0.1
msgpack==1.1.1
multidict==6.7.0
packaging==25.0
pillow==11.3.0
postgrest==2.23.0
propcache==0.4.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
pyOpenSSL==25.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
realtime==2.23.0
redis==6.4.0
reportlab==4.2.5
requests==2.32.5
s3transfer==0.10.4
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
storage3==2.23.0
StrEnum==0.4.15
stripe==11.1.0
supabase==2.23.0
supabase-auth==2.23.0
supabase-functions==2.23.0
tomli==2.3.0
Twisted==25.5.0
txaio==25.6.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
whitenoise==6.10.0
yarl==1.22.0
zope.interface==8.0
